[*] rename sources.MonitoredDatabase to SourceConn Also move some connection specific functions from reaper to SourceConn methods, e.g. TryDiscoverExecutionEnv(), DoesFunctionExists()
